[master] Merge q4-working-branch to master (#3780)
* Made modifications to logger to make LOG_MARKER very quiet
* Changed flag in lookup.cpp to send everything to dsgaurd
* Replaced isolated-server-accounts.json with single entry for use with the scripts:
   FundFromPrimaryAccount.ts in the hardhat EvmAcceptanceTests script dir
* Updated constants.xml
* reduced extra time for consensus as described in Zelensky doc.
* localdev.py now defaults to using isolated-server-accounts
* ZIL-5326: remove dependency on Python3 (#3742)
* Use Boost.Process to run download_incr_DB.py with python3.
* Switch boost::filesystem -> std::filesystem.
* Experimental use of vcpkg binary caching in Dockerfile.
* p2pcomm refactor pt.1
* Sending part, relaxed blacklisting policy
* Don't assign empty pubkeys
* Ooutgoing connections to multiplier nodes follow the old protocol (w/reconnects)
* Remove dependency on libevent.
* Upgraded Blacklist to to be keyed by ip/port/node combo.
   This is only a partial implementation, need to do the same for whitelist and push through rest of arch.
   This change needs to be done in such a way that preserves existing functionality at the ip level.
* Fix reciept output as per spec (#3759)
* Fix codehash to get only contract code hash (#3744) (#3763)
* Modify PoW submission for ds only diff (#3765)
* Fix tests
* Small adjustments and removal of unwanted part of the protocol, originally introduced to overcome limitations of godupliactor in release 8.x (#3786)
* Change m_noKill -> m_kill in daemon.
* Remove dead code

---------

+++ daemon/ZilliqaDaemon.cpp
@@ -68,7 +68,8 @@ ZilliqaDaemon::ZilliqaDaemon(int argc, const char* argv[], std::ofstream& log)
       m_recovery(0),
       m_nodeIndex(0),
       m_syncType(0),
-      m_cseed(false) {
+      m_cseed(false),
+      m_kill(true) {
   if (ReadInputs(argc, argv) != SUCCESS) {
     ZilliqaDaemon::LOG(m_log, "Failed to read inputs.");
     exit(EXIT_FAILURE);
@@ -138,13 +139,13 @@ void ZilliqaDaemon::MonitorProcess(const string& name,
   for (const pid_t& pid : pids) {
     // If sig is 0 (the null signal), error checking is performed but no signal
     // is actually sent
-    if (kill(pid, 0) < 0) {
+    if (m_kill and kill(pid, 0) < 0) {
       if (errno == EPERM) {
         ZilliqaDaemon::LOG(m_log, "Daemon does not have permission Name: " +
                                       name + " Id: " + to_string(pid));
       } else if (errno == ESRCH) {
-        ZilliqaDaemon::LOG(
-            m_log, "Process died Name: " + name + " Id: " + to_string(pid));
+        ZilliqaDaemon::LOG(m_log, "We think Process died Name: " + name +
+                                      " Id: " + to_string(pid));
         m_died[pid] = true;
       } else {
         ZilliqaDaemon::LOG(m_log, "Kill failed due to " + to_string(errno) +
@@ -397,8 +398,13 @@ void ZilliqaDaemon::StartScripts() {
 
   string cmdToRun = "ps axf | grep " + script +
                     " | grep -v grep  | awk '{print \"kill -9 \" $1}'| sh &";
-  ZilliqaDaemon::LOG(m_log, "Start to run command: \"" + cmdToRun + "\"");
-  ZilliqaDaemon::LOG(m_log, "\" " + Execute(cmdToRun + " 2>&1") + " \"");
+
+  if (m_kill) {
+    ZilliqaDaemon::LOG(m_log, "Start to run command: \"" + cmdToRun + "\"");
+    ZilliqaDaemon::LOG(m_log, "\" " + Execute(cmdToRun + " 2>&1") + " \"");
+  } else {
+    ZilliqaDaemon::LOG(m_log, "Not running command: \"" + cmdToRun + "\"");
+  }
 
   cmdToRun = "python3 " + m_curPath + script +
              (0 == m_nodeIndex ? "" : " -f 10") + " &";
@@ -408,13 +414,11 @@ void ZilliqaDaemon::StartScripts() {
   Exit(0);
 }
 
-void ZilliqaDaemon::Exit(int exitCode)
-{
+void ZilliqaDaemon::Exit(int exitCode) {
   // Since the updater uses the Logger and the daemon keeps fork-ing
   // we can't realy on exit() because it will hang when the logger
   // tries to shutdown (since the child won't have the same running threads).
-  if (m_updater)
-  {
+  if (m_updater) {
     m_log.flush();
     _exit(exitCode);
   }
@@ -425,10 +429,13 @@ void ZilliqaDaemon::Exit(int exitCode)
 void ZilliqaDaemon::KillProcess(const string& procName) {
   vector<pid_t> pids = ZilliqaDaemon::GetProcIdByName(procName);
   for (const auto& pid : pids) {
-    ZilliqaDaemon::LOG(
-        m_log, "Killing " + procName + " process before launching daemon...");
-    kill(pid, SIGTERM);
-    ZilliqaDaemon::LOG(m_log, procName + " process killed successfully.");
+    if (m_kill) {
+      ZilliqaDaemon::LOG(
+          m_log, "Killing " + procName + " process before launching daemon...");
+
+      kill(pid, SIGTERM);
+      ZilliqaDaemon::LOG(m_log, procName + " process killed successfully.");
+    }
   }
 }
 
@@ -453,7 +460,8 @@ int ZilliqaDaemon::ReadInputs(int argc, const char* argv[]) {
       "logpath,g", po::value<string>(&m_logPath),
       "customized log path, could be relative path (e.g., \"./logs/\"), or "
       "absolute path (e.g., \"/usr/local/test/logs/\")")(
-      "cseed,c", "Runs as cummunity seed node if set");
+      "cseed,c", "Runs as cummunity seed node if set")(
+      "killnone,k", "does not kill processes");
 
   po::variables_map vm;
 
@@ -473,6 +481,11 @@ int ZilliqaDaemon::ReadInputs(int argc, const char* argv[]) {
     if (vm.count("cseed")) {
       ZilliqaDaemon::LOG(m_log, "Running Daemon for community seed node.");
       m_cseed = true;
+    }
+    if (vm.count("killnone")) {
+      ZilliqaDaemon::LOG(
+          m_log, "does not kill things - useful for experimental native.");
+      m_kill = false;
     }
   } catch (boost::program_options::required_option& e) {
     ZilliqaDaemon::LOG(m_log, "ERROR: " + string(e.what()));
@@ -528,6 +541,7 @@ int main(int argc, const char* argv[]) {
   bool startNewByDaemon = true;
   while (1) {
     for (const auto& name : programName) {
+      std::cout << "Monitoring " << name << " process..." << std::endl;
       daemon.MonitorProcess(name, startNewByDaemon);
     }
